+++ Tests/Realm.Tests/Database/TestObjects.cs
@@ -799,7 +799,7 @@ namespace Realms.Tests
 
         public int Age { get; set; }
 
-        [Backlink(nameof(Owner.Dogs))]
+        [Backlink(nameof(Owner.ListOfDogs))]
         public IQueryable<Owner> Owners { get; }
     }
 
@@ -809,7 +809,9 @@ namespace Realms.Tests
 
         public Dog TopDog { get; set; }
 
-        public IList<Dog> Dogs { get; }
+        public IList<Dog> ListOfDogs { get; }
+
+        public ISet<Dog> SetOfDogs { get; }
     }
 
     // A copy of Owner that verifies that different objects referring to the same type (Dog)
@@ -820,7 +822,9 @@ namespace Realms.Tests
 
         public Dog TopDog { get; set; }
 
-        public IList<Dog> Dogs { get; }
+        public IList<Dog> ListOfDogs { get; }
+
+        public ISet<Dog> SetOfDogs { get; }
     }
 
     public class RequiredStringObject : RealmObject